Add two new options: --iso-host --iso-path
iso-host is the server name that is hosting the ISO, e.g. yow-myousaf-lx.wrs.com.  iso-path is the full location of the ISO, e.g. /home/myousaf/mybootimage.iso.

Both parameters must be specified.  Adding --burn-usb is optional, as once both --iso-host and --iso-path are provided, burn_usb is assumed.