US104439: pike rebase cinderclient: cinder-show no longer show attachments
This fix 2 issues:
1. Pike version of cinder-show command no longer show "attachments" field
and this broke sanity test scripts expecting this information.

2. The expection in test_evacuate_vms() for live migration is not
correct.

From Chris Friesen:
In the case of boot-from-image with LVM storage on the compute we will
fail live migration regardless of the ephemeral value.

In the case of boot-from-volume with LVM storage on the compute we might
succeed if both ephemeral and swap are 0 (since we're only using the
cinder volume), and fail if either is non-zero."

With this knowledge, vm3 and vm4 would fail live-migration.

Remove vm3 and vm4 from vms_to_mig list for now. In the future, TC
test_evacuate_vms() could be updated to launch these vms on the target
host directly in order to have a good coverage on evacuation.